# bff-server/swagger.yaml
openapi: 3.0.3
info:
  title: Moodle BFF API
  description: Backend For Frontend API for Moodle integration
  version: 1.0.0
  contact:
    name: Moodle BFF API Support
servers:
  - url: https://52.194.117.196
    description: Production server (HTTPS via nginx)
  - url: http://52.194.117.196
    description: Development server (HTTP, redirects to HTTPS)
  - url: http://localhost:3001
    description: Local development (direct access)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Cognito JWT認証を使用します。

        1. Cognitoでログインして IDToken を取得
        2. "Authorize" ボタンをクリック
        3. IDToken を入力（"Bearer" プレフィックスは不要）

        トークン取得方法:
        ```bash
        # AWS CLIで取得
        aws cognito-idp initiate-auth \
          --auth-flow USER_PASSWORD_AUTH \
          --client-id 23jacbr6nk4baiftjueddmr4kb \
          --auth-parameters USERNAME=your-email,PASSWORD=your-password \
          --region ap-northeast-1

        # またはスクリプトを使用
        ./get-token.sh your-email your-password
        node test-auth.js your-email your-password
        ```
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
...
    UserInfo:
      type: object
      properties:
        userid:
          type: integer
        username:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        fullname:
          type: string
        email:
          type: string
        userpictureurl:
          type: string
...
paths:
...
  /api/user/info:
    get:
      tags:
        - Authentication
      summary: Get current user info
      description: Get information about the currently authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User information retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/coaching/coaches/{coach_user_id}/students:
    get:
      tags:
        - Coaching
      summary: コーチの受講生一覧取得
      description: コーチが担当する受講生のIDリストを取得します（管理者またはコーチ本人のみ）
      security:
        - bearerAuth: []
      parameters:
        - name: coach_user_id
          in: path
          required: true
          schema:
            type: integer
          description: コーチのMoodleユーザーID
        - name: include_deleted
          in: query
          schema:
            type: boolean
            default: false
          description: 削除済みも含めるか
      responses:
        '200':
          description: 受講生一覧取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  coach_user_id:
                    type: integer
                  student_user_ids:
                    type: array
                    items:
                      type: integer
        '403':
          description: Forbidden
  /api/coaching/students/{student_user_id}/coach:
    get:
      tags:
        - Coaching
      summary: 受講生のコーチ取得
      description: 受講生に割り当てられたコーチのIDを取得します（管理者、コーチ、または学生本人のみ）
      security:
        - bearerAuth: []
      parameters:
        - name: student_user_id
          in: path
          required: true
          schema:
            type: integer
          description: 受講生のMoodleユーザーID
        - name: include_deleted
          in: query
          schema:
            type: boolean
            default: false
          description: 削除済みも含めるか
      responses:
        '200':
          description: コーチ情報取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  student_user_id:
                    type: integer
                  coach_user_id:
                    type: integer
                    nullable: true
        '403':
          description: Forbidden
...